Consolidate ViewMenu into ViewManager
