+++ desktop/cmp/viewmanager/ViewManager.scss
@@ -1,12 +1,35 @@
 .xh-view-manager {
   align-items: center;
 
-  // Save Button
-  & > .xh-button {
-    margin-right: var(--xh-pad-half-px);
+  &__menu-item {
+    &:hover {
+      .xh-view-manager__menu-item__fave-toggle {
+        opacity: 1;
+        color: var(--xh-yellow);
+      }
+    }
+
+    &__fave-toggle {
+      opacity: 0;
+
+      &:hover {
+        color: var(--xh-yellow-light) !important;
+      }
+
+      &--active {
+        opacity: 1;
+        color: var(--xh-yellow);
+      }
+
+      &--active:hover {
+        color: var(--xh-yellow-light) !important;
+      }
+    }
   }
 
+  //------------------------
   // Dialogs
+  //------------------------
   &__manage-dialog,
   &__save-dialog {
     &__form {
@@ -30,32 +53,6 @@
     &__metadata {
       color: var(--xh-text-color-muted);
       font-size: var(--xh-font-size-small-px);
-    }
-  }
-
-  &__menu-item {
-    &:hover {
-      .xh-view-manager__menu-item--fav {
-        opacity: 1;
-        color: var(--xh-yellow);
-      }
-    }
-
-    &--fav {
-      opacity: 0;
-
-      &:hover {
-        color: var(--xh-yellow-light) !important;
-      }
-
-      &--active {
-        opacity: 1;
-        color: var(--xh-yellow);
-      }
-
-      &--active:hover {
-        color: var(--xh-yellow-light) !important;
-      }
     }
   }
 }
