+++ desktop/cmp/viewmanager/index.ts
@@ -1,4 +1,3 @@
 export * from './ViewManager';
 export * from './cmp/ManageDialog';
 export * from './cmp/SaveDialog';
-export * from './cmp/ViewMenu';
